[STF] Add non-host sharing test
The new `EditDuringNonHostAddsProjectTest` is derived from the old and disabled
`NonHostAddsProjectsWithHostDelayTest`. The latest non-host-project sharing
implementation does not allow the delay anymore (as the project is at first
only shared with the host), but reversing the order of project acceptance of the
old test gives a good consistency tests with multiple edits during the negotiations
of all testers.
